#112 feat: form detail 관련 컴포넌트 및 페이지 폴더 구조 변경
