+++ src/components/CoachLecture/FormDetail/PlayerInfo/PlayerInfo.tsx
