+++ src/components/CoachLecture/FormDetail/PlayerInfo/PlayerInfoStyle.ts
